Added four ncurses 'extended' color functions (extended_color_content(), extended_pair_content(), init_extended_color(), init_extended_pair()).  No real change in underlying functionality... yet,  but this is a necessary first step.

+++ curspriv.h
@@ -48,7 +48,7 @@ typedef struct           /* structure for ripped off lines */
 
 void    PDC_beep(void);
 bool    PDC_can_change_color(void);
-int     PDC_color_content(short, short *, short *, short *);
+int     PDC_color_content(int, int *, int *, int *);
 bool    PDC_check_key(void);
 int     PDC_curs_set(int);
 void    PDC_doupdate(void);
