+++ sdl1/pdcscrn.c
@@ -342,7 +342,7 @@ bool PDC_can_change_color(void)
     return TRUE;
 }
 
-int PDC_color_content(short color, short *red, short *green, short *blue)
+int PDC_color_content(int color, int *red, int *green, int *blue)
 {
     *red = DIVROUND(pdc_color[color].r * 1000, 255);
     *green = DIVROUND(pdc_color[color].g * 1000, 255);
