+++ vt/pdcscrn.c
@@ -374,7 +374,7 @@ bool PDC_can_change_color(void)
     return TRUE;
 }
 
-int PDC_color_content( short color, short *red, short *green, short *blue)
+int PDC_color_content( int color, int *red, int *green, int *blue)
 {
     const PACKED_RGB col = PDC_get_palette_entry( color);
