+++ x11/pdcscrn.c
@@ -701,7 +701,7 @@ bool PDC_can_change_color(void)
     return TRUE;
 }
 
-int PDC_color_content(short color, short *red, short *green, short *blue)
+int PDC_color_content(int color, int *red, int *green, int *blue)
 {
     XColor tmp;
     Colormap cmap = DefaultColormap(XCURSESDISPLAY,
